Fix function overloading (#136)
Previously, function overloading didn't call the mapped class' function,
but instead just assigned the type. This results in an overwritten
function e.g., not casting the return type as a struct for a specific
overload. We now build a function signature per overloaded function, and
do a lookup when doing `__call__` when passed arguments in.

This feature is critical for error handling in `call` and `transact`, as
we need to provide specific information based on which contract function
gets called.

Major changes:
- Generating function signature, and looking up said signature when
accessing a specific function overload.
- Removing case statement in jinja for generating separate code if the
function doesn't have overloads. Instead, a function without overloads
is treated as an overloaded function with one overload.
- Adding support for `kwargs` when specifying function args.
- Adding support for vector of struct input and outputs for function
overloads.